untangle: unix_timestamp lives with the identity records
access_policy stamps created_at/revoked_at on identity records but
reached back into peer/pairing for the clock helper -- the last
non-test access->peer ref after the access_policy move. The helper
moves to access_policy (records are its domain); pairing and
access_request now point downward at it, the direction the layering
wants.
